backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
supabase==2.10.0
httpx==0.27.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
aiohttp==3.11.4
beautifulsoup4==4.12.3
tldextract==5.1.3
stripe==11.2.0
python-dotenv==1.0.1
